Add support for libunwind-based stack unwinding
libunwind is a portable, lightweight, widely adopted library to unwind the stack.
It is part of GCC via libiberty, and LLVM as well.

It is an alternative to the `backtrace` approach, which is not always available,
e.g. Musl libc does not provide a `backtrace` implementation, and the alternative,
libexecinfo, is known to have some bugs.

This is currently hidden behind a version, meaning only packagers or power users
will get access to it, so it can be properly tested.